+++ libraries/provider_dnsmasq_local.rb
@@ -52,7 +52,7 @@ class Chef
           notifies :restart, "dnsmasq_local_service[#{new_resource.name}]"
         end
         dnsmasq_local_service new_resource.name do
-          new_resource.environment.each { |k, v| send(k, v) }
+          new_resource.options.each { |k, v| send(k, v) }
           notifies :restart, "dnsmasq_local_service[#{new_resource.name}]"
         end
       end