+++ libraries/provider_dnsmasq_local_service_debian.rb
@@ -30,31 +30,6 @@ class Chef
       if defined?(provides)
         provides :dnsmasq_local_service, platform_family: 'debian'
       end
-
-      #
-      # Generate the `/etc/default/dnsmasq` file for the service.
-      #
-      action :create do
-        merged_env = new_resource.environment.merge(
-          new_resource.state.select { |k, v| k != :environment && !v.nil? }
-        )
-        file '/etc/default/dnsmasq' do
-          header = <<-EOH.gsub(/^ +/, '')
-            # This file is managed by Chef.
-            # Any changes to it will be overwritten.
-          EOH
-          body = Hash[merged_env.sort].map { |k, v| "#{k.upcase}='#{v}'" }
-                 .join("\n")
-          content(header + body)
-        end
-      end
-
-      #
-      # Clean up the service files that are managed by Chef.
-      #
-      action :remove do
-        file('/etc/default/dnsmasq') { action :delete }
-      end
     end
   end
 end