+++ libraries/resource_dnsmasq_local.rb
@@ -36,9 +36,9 @@ class Chef
       attribute :config, kind_of: Hash, default: {}
 
       #
-      # Supporting passing dnsmasq environment variables as a hash.
+      # Support passing command line options for dnsmasq as a hash.
       #
-      attribute :environment, kind_of: Hash, default: {}
+      attribute :options, kind_of: Hash, default: {}
     end
   end
 end