+++ spec/resources/dnsmasq_local_app/rhel.rb
@@ -15,10 +15,8 @@ shared_context 'resources::dnsmasq_local_app::rhel' do
     context 'the :remove action' do
       include_context description
 
-      shared_examples_for 'any RHEL platform' do
-        it 'removes the dnsmasq package' do
-          expect(chef_run).to remove_package('dnsmasq')
-        end
+      it 'removes the dnsmasq package' do
+        expect(chef_run).to remove_package('dnsmasq')
       end
     end
   end