+++ test/integration/custom/serverspec/localhost/service_spec.rb
@@ -8,9 +8,7 @@ describe 'dnsmasq-local::custom::service' do
       expected = <<-EOH.gsub(/^ {8}/, '').strip
         # This file is managed by Chef.
         # Any changes to it will be overwritten.
-        CONFIG_DIR='/etc/dnsmasq.d,.dpkg-dist,.dpkg-old,.dpkg-new'
         DNSMASQ_OPTS='--bind-dynamic'
-        ENABLED='1'
       EOH
       expect(subject.content).to eq(expected)
     end